Fix for uncalled service callback for App2AppService. (#14)
Between each android application instance, its resources get released and all running services are stopped. When each service is stopped, a callback is called releasing more resources if needed. For App2AppService, its callback was not being called, and as a result the subsequent applications could not instantiate an App2AppService object.

The problem was in the App2AppLocalService::OnRemoteServiceStopped(), which called the overrided Stop() method and at its current state, no callback was being called.

Co-authored-by: Giorgos Epitropou <giorgos.epitropou@oceanbluesoftware.co.uk>