Split period changed event from stream updated event on dash proxy.
Dash proxy used to dispatch "__obs_onperiodchanged__" event when "periodSwitchCompleted" and "streamUpdated" events were dispatched. This led to undefined behavior as the "streamUpdated" event is dispatched multiple times when the manifest has more than one periods.
This was noticed with media synchroniser at the time where it was supposed to switch timeline monitoring when the stream changes period.

Introduced additional event "__obs_onstreamupdated__", which media synchroniser will use to stop timeline monitoring when a previously monitored period is no longer present in the manifest.