Pass org.hbbtv_HTML5-DASH003 (#324)
HBBTV 9.4.2
The origin of the media timeline used by HTML5 media elements (i.e. the
&lt;audio&gt; and &lt;video&gt; elements) shall be the start time of the
first Period that was defined in the MPD when the MPD was first loaded.
The origin of the media timeline shall not change unless the HTML5 media
element source is changed or the load() method is called. [...] For a
dynamic MPD, getStartDate() shall return a Date oject representing the
value of MPD@availabilityStartTime plus the PeriodStart time (see clause
5.3.2.1 of MPEG DASH ISO/IEC 23009-1 [29]) of the first regular Period
when the MPD was first loaded.

Thus:
- startDate is only calculated once.
- startDate is forgotten on load(calls).

In addition to that, the loadeddata event is not triggered for this MPD
(first period has a start of 3 seconds, dashjs bug?), so we also watch
for progress events to call play() when play() was called immediately
after setting the src attribute. progress events are fired with this
MPD.