FREE-128 jenkins: create orb auto roller
Add a Jenkins pipeline that updates the DEPS file of the Chromium repo
with the latest ORB repo commit (in the Chromium project this process
is known as "rolling").

NB: If both Chromium and ORB are branched and an auto roller is setup
for the branch, the parameter DEPS_REPO_BRANCH within the Jenkinsfile of
the ORB branch will need to be updated with the correct Chromium repo
branch to change.