FREE-273 OpApp and HbbTVApp use common BaseApp class.
Replace the unordered_map with OpApp and HbbTVApp members (unique_ptrs). This changes the manager quite a bit but is much cleaner and a better way to separate out concerns.

The BaseApp is reponsible for holding type and state. State enumeration is really only applicable to OpApp, so may change in future.

Separate runOpApp and runHbbTVApp methods.

TO DO: Fix sessionCallback vector.