FREE-273 Remove the SetCurrentInterface method.
The ApplicationSessionCallback is registered against an apptype and since app can be recovered the correct session callback can also be dereferenced.
Previously solution was brittle and required explicitly setting the interface.

Defined APP_TYPE_MAX as a more robust check limit.

Note the multi-element list of callbacks is still not ideal and should be fixed at a later date. See FREE-298