FREE-285 Application Manager refactoring (#421)
* FREE-273 Refactoring for readability:

Make CreateAndRunApp public for testing.
Remove verbose and redundant method descriptions.
Private methods start lowercase.

* FREE-273 Removing duplicate method descriptions from implementation.

* FREE-273 Adding CreateAndRunApp unit tests.

* FREE-273 Replaces map object with explicit members for HbbTV and OpApp.

* FREE-273 OpApp and HbbTVApp use common BaseApp class.

Replace the unordered_map with OpApp and HbbTVApp members (unique_ptrs). This changes the manager quite a bit but is much cleaner and a better way to separate out concerns.

The BaseApp is reponsible for holding type and state. State enumeration is really only applicable to OpApp, so may change in future.

Separate runOpApp and runHbbTVApp methods.

TO DO: Fix sessionCallback vector.

* FREE-273 Removing unnecessary definition of app type enum.

Adds a GetApplicationType to IAppMgrInterface to remove need for storing it in moderator (which only needs to store interface to application manager).

* FREE-273 Remove the SetCurrentInterface method.

The ApplicationSessionCallback is registered against an apptype and since app can be recovered the correct session callback can also be dereferenced.
Previously solution was brittle and required explicitly setting the interface.

Defined APP_TYPE_MAX as a more robust check limit.

Note the multi-element list of callbacks is still not ideal and should be fixed at a later date. See FREE-298

* FREE-273 Fixing missing return.